[upd] Pre-publication
