+++ calculations.py
@@ -7,28 +7,28 @@ sbatch_template = '''#! /bin/bash
 #SBATCH --job-name="NN_{functional}"
 #SBATCH --gpus=1
 #SBATCH --nodes=1
-#SBATCH --cpus-per-task=4
+#SBATCH --cpus-per-task=2
 #SBATCH --gpus-per-node=1
 #SBATCH --mail-user=schneider.mark14@gmail.com
 #SBATCH --mail-type=ALL
-#SBATCH --output="/home/mmedvedev/schnm/log/22_JULY/MSE_polusumma/bs_2_no_ae17/{functional}_0.4_{omega:.5f}_"%j.out
+#SBATCH --output="/home/mmedvedev/schnm/log/13_august/RELATIVE_AE17/{functional}_0.4_{omega:.5f}_"%j.out
 #SBATCH --constraint="type_e"
 #SBATCH --time=3-0
 # Executable
-python test.py --Name {functional} --N_preopt 3 --N_train 500 --Batch_size {batch_size} --Dropout 0.6 --Omega {omega:.5f} --LR_predopt 0.02'''
+python test.py --Name {functional} --N_preopt 3 --N_train 300 --Batch_size {batch_size} --Dropout 0.6 --Omega {omega:.5f} --LR_predopt 0.02'''
 
+
+# Names and batch sizes
 functionals = [
-    ("PBE_8_32", 5),
-#    ("XALPHA_32_32", 6),
-#    ("XALPHA_16_32", 12),
-#    ("XALPHA_8_32", 12),
-#    ("XALPHA_4_512", 1)
+    ("PBE_8_32", 4),
+    ("XALPHA_32_32", 6),
+    ("XALPHA_16_32", 12),
+    ("XALPHA_8_32", 12),
+    ("XALPHA_4_512", 1)
 ]
 
 n = 9
 omegas = list(np.roots(chop(taylor(lambda x: chebyt(n, x), 0, n))[::-1])/2+0.5) + [0,]
-
-#omegas = [0,]
 
 for functional, batch_size in functionals:
     for omega in omegas:
