+++ dft_functionals/SVWN3.py
@@ -1,4 +1,5 @@
 import torch
+import numpy as np
 
 from utils import catch_nan
 
@@ -141,7 +142,7 @@ def f_vwn(rs, z, c_arr):
 
 def rs_z_calc(rho):
     eps = 1e-29
-    rs = (3 / ((rho[:, 0] + rho[:, 1] + eps) * (4 * torch.pi))) ** (1 / 3)
+    rs = (3 / ((rho[:, 0] + rho[:, 1] + eps) * (4 * np.pi))) ** (1 / 3)
     z = (rho[:, 0] - rho[:, 1]) / (rho[:, 0] + rho[:, 1] + eps)
     catch_nan(rs=rs, z=z)
     return rs, z
@@ -149,8 +150,8 @@ def rs_z_calc(rho):
 
 # SLATER
 
-LDA_X_FACTOR = -3 / 8 * (3 / torch.pi) ** (1 / 3) * 4 ** (2 / 3)  # param
-RS_FACTOR = (3 / (4 * torch.pi)) ** (1 / 3)
+LDA_X_FACTOR = -3 / 8 * (3 / np.pi) ** (1 / 3) * 4 ** (2 / 3)  # param
+RS_FACTOR = (3 / (4 * np.pi)) ** (1 / 3)
 DIMENSIONS = 3
