+++ predopt.py
@@ -33,7 +33,7 @@ true_constants_PBE = torch.Tensor(
     [
         [
             0.06672455060314922,
-            (1 - torch.log(torch.Tensor([2]))) / (torch.pi**2),
+            (1 - torch.log(torch.Tensor([2]))) / (np.pi**2),
             1.709921,
             7.5957,
             14.1189,
@@ -54,7 +54,7 @@ true_constants_PBE = torch.Tensor(
             0.21370,
             0.20548,
             0.11125,
-            -3 / 8 * (3 / torch.pi) ** (1 / 3) * 4 ** (2 / 3),
+            -3 / 8 * (3 / np.pi) ** (1 / 3) * 4 ** (2 / 3),
             0.8040,
             0.2195149727645171,
         ]
