+++ predopt_train.py
@@ -2,37 +2,135 @@ import gc
 import pickle
 from optparse import OptionParser
 
-import mlflow
 import numpy as np
 import torch
 from torch import nn
 from tqdm.notebook import tqdm
+import os
 
 from dataset import collate_fn, collate_fn_predopt
-from NN_models import MLOptimizer
+from NN_models import MLOptimizer, pcPBEMLOptimizer
 from predopt import DatasetPredopt, true_constants_PBE
 from prepare_data import load_chk
 from reaction_energy_calculation import (calculate_reaction_energy,
                                          get_local_energies)
-from utils import log_params, set_random_seed
+from utils import set_random_seed
+import matplotlib.pyplot as plt
+from torch.optim.lr_scheduler import StepLR
+
 
 set_random_seed(41)
 
 
-MEAN_TRAIN_FEATURES = torch.tensor(
-    [
-        -6.33887041,
-        -6.68211794,
-        -8.682011,
-        -7.72950894,
-        -8.96326543,
-        -6.54985721,
-        -6.87318032,
-    ]
-)
-STD_TRAIN_FEATURES = torch.tensor(
-    [5.00045545, 5.19493982, 6.30814192, 6.72813273, 6.34910604, 5.24995811, 5.41875425]
-)
+FCHEM_VALIDATION = {
+    'MGAE109': 1/4.73394495412844,
+    'NCCE31': 10,
+#    'AE17': 1/100
+}
+
+TRAIN_WEIGHTS = {
+    "ABDE4": 1/93.42/4,
+    "AE17": 1/93753.82/17,
+    "DBH76": 1/18.40/76,
+    "EA13": 1/38.31/13,
+    "IP13": 1/253.83/13,
+    "MGAE109": 1/498.53/109,
+    "NCCE31": 1/3.63/31,
+    "PA8": 1/163.33/8,
+    "pTC13": 1/168.79/13,
+}
+
+mean_train_weight = np.mean(np.array(list(TRAIN_WEIGHTS.values())))
+
+PBE_TRAIN_ERRORS = {
+    "ABDE4": 2.347,
+    "AE17": 54.272,
+    "DBH76": 10.120,
+    "EA13": 3.112,
+    "IP13": 4.964,
+    "MGAE109": 4.041,
+    "NCCE31": 16.937,
+    "PA8": 1.688,
+    "pTC13": 5.784,
+}
+
+PBE_VALIDATION_ERRORS = {
+    "ABDE4": 5.689,
+    "AE17": 66.095,
+    "DBH76": 8.904,
+    "EA13": 2.451,
+    "IP13": 3.358,
+    "MGAE109": 3.974,
+    "NCCE31": 15.767,
+    "PA8": 1.789,
+    "pTC13": 10.985,
+}
+
+
+def loss_function(factor_dictionary, total_database_errors, val=False):
+    fchem = 0
+
+    for db in sorted(total_database_errors):
+        factor = factor_dictionary.get(db, 1)
+        error = np.sqrt((np.mean(np.array(total_database_errors[db])**2)))
+        fchem += error*factor
+        if val:
+            err_dict = PBE_VALIDATION_ERRORS
+        else:
+            err_dict = PBE_TRAIN_ERRORS
+        print(f'{db}: {error*factor:6.3f}, {100*(error*factor-err_dict.get(db))/err_dict.get(db):6.3f}%')
+
+    return fchem
+
+
+def batch_fchem(current_bases, reaction_energy, y_batch):
+    mse = nn.MSELoss()
+    err_dict = dict()
+    errors = []
+
+
+    for database, pred, ref in zip(current_bases, reaction_energy, y_batch):
+        err_dict[database] = err_dict.get(database, [[], []])
+        err_dict[database][0].append(pred)
+        err_dict[database][1].append(ref)
+    for database in err_dict:
+        if database == 'AE17':
+            factor = np.abs((1/err_dict[database][1][-1]).item())/mean_train_weight
+        else:
+            factor = TRAIN_WEIGHTS.get(database, 1)/mean_train_weight
+        errors.append(factor*mse(torch.stack(err_dict[database][0]), torch.stack(err_dict[database][1])))
+
+    fchem = torch.sqrt(1e-20 + torch.sum(torch.stack(errors)))/len(err_dict)
+    return fchem
+    
+
+
+def extend_bases(X_batch, bases):
+    if len(X_batch["Database"][0]) == 1:
+        current_bases = [X_batch["Database"],]
+    else:
+        current_bases = list(X_batch["Database"])
+    bases += current_bases
+
+    return current_bases, bases
+
+
+def make_total_db_errors(pred_energies, reaction_energy, errors, ref_energies, y_batch, total_database_errors, current_bases):
+    if len(pred_energies): 
+        pred_energies = torch.hstack([pred_energies, reaction_energy])
+        ref_energies = torch.hstack([ref_energies, y_batch])
+        errors = torch.hstack([errors, reaction_energy-y_batch])
+    else:
+        pred_energies = reaction_energy
+        ref_energies = y_batch
+        errors = reaction_energy-y_batch
+
+    for base, error in zip(current_bases, reaction_energy-y_batch):
+        total_database_errors[base] = total_database_errors.get(base, [])
+        total_database_errors[base].append((torch.abs(error).item()))
+
+    return pred_energies, ref_energies, errors, total_database_errors
+
 
 
 data, data_train, data_test = load_chk(path="checkpoints")
@@ -42,52 +140,66 @@ parser.add_option('--Name', type=str,
                   help="Name of the functional",
                   default="PBE_8_32")
 parser.add_option('--N_preopt', type=int,
-                  default=50,
+                  default=3,
                   help="Number of pre-optimization epochs")
 parser.add_option('--N_train', type=int,
-                  default=184,
+                  default=1000,
                   help="Number of training epochs")
 parser.add_option('--Batch_size', type=int,
-                  default=8,
+                  default=3,
                   help="Number of reactions in a batch")
 parser.add_option('--Dropout', type=float,
-                  default=0.4,
+                  default=0.6,
                   help="Dropout rate during training")
 parser.add_option('--Omega', type=float,
-                  default=0.0412,
+                  default=0.0,
+                  help="Omega value in the loss function")
+parser.add_option('--LR_train', type=float,
+                  default=1e-4,
+                  help="Omega value in the loss function")
+parser.add_option('--LR_predopt', type=float,
+                  default=2e-2,
                   help="Omega value in the loss function")
 
 (Opts, args) = parser.parse_args()
 
-name, n_predopt, n_train, batch_size, dropout, omega = (
+name, n_predopt, n_train, batch_size, dropout, omega, lr_train, lr_predopt = (
     Opts.Name,
     Opts.N_preopt,
     Opts.N_train,
     Opts.Batch_size,
     Opts.Dropout,
     Opts.Omega,
+    Opts.LR_train,
+    Opts.LR_predopt
 )
+
+print('name, n_predopt, n_train, batch_size, dropout, omega, lr_train, lr_predopt')
+print(name, n_predopt, n_train, batch_size, dropout, omega, lr_train, lr_predopt)
+
 
 if "PBE" in name:
     rung = "GGA"
     dft = "PBE"
-    nconstants = 24
-    lr_predopt = 3e-3
-    lr_train = 3e-4
+
 elif "XALPHA" in name:
     rung = "LDA"
     dft = "XALPHA"
     nconstants = 1
-    lr_predopt = 6e-3
-    lr_train = 2e-3
+
 
 num_layers, h_dim = map(int, name.split("_")[1:])
 device = torch.device("cuda:0") if torch.cuda.is_available else torch.device("cpu")
-constants = [0,1,22,23]
-#constants = []
-model = MLOptimizer(
-    num_layers=num_layers, h_dim=h_dim, nconstants=nconstants, dropout=dropout, DFT=dft, constants=constants
-).to(device)
+
+
+if dft == "PBE":
+    model = pcPBEMLOptimizer(
+        num_layers=num_layers, h_dim=h_dim, dropout=dropout, DFT=dft
+    ).to(device)
+elif dft == "XALPHA":
+    model = MLOptimizer(
+        num_layers, h_dim, nconstants, dropout, dft
+    ).to(device)
 
 # Load dispersion corrections.
 with open("./dispersions/dispersions.pickle", "rb") as handle:
@@ -99,7 +211,6 @@ class Dataset(torch.utils.data.Dataset):
         self.data = data
 
     def __getitem__(self, i):
-        self.data[i].pop("Database", None)
         return self.data[i], self.data[i]["Energy"]
 
     def __len__(self):
@@ -110,7 +221,7 @@ train_set = Dataset(data=data_train)
 train_dataloader = torch.utils.data.DataLoader(
     train_set,
     batch_size=batch_size,
-    num_workers=4,
+    num_workers=2,
     pin_memory=True,
     shuffle=True,
     collate_fn=collate_fn,
@@ -119,7 +230,7 @@ test_set = Dataset(data=data_test)
 test_dataloader = torch.utils.data.DataLoader(
     test_set,
     batch_size=batch_size,
-    num_workers=4,
+    num_workers=2,
     pin_memory=True,
     shuffle=False,
     collate_fn=collate_fn,
@@ -128,15 +239,15 @@ train_predopt_set = DatasetPredopt(data=data, dft=dft)
 train_predopt_dataloader = torch.utils.data.DataLoader(
     train_predopt_set,
     batch_size=batch_size,
-    num_workers=4,
+    num_workers=2,
     pin_memory=True,
-    shuffle=True,
+    shuffle=False,
     collate_fn=collate_fn_predopt,
 )
 
 criterion = nn.MSELoss()
 
-mlflow.set_experiment(name)
+
 name += "_" + str(dropout)
 
 optimizer = torch.optim.Adam(
@@ -167,18 +278,21 @@ train_loss_mse, train_loss_mae = predopt(
     n_epochs=n_predopt,
     accum_iter=1,
 )
-log_params(model, train_loss_mse, train_loss_mae, name=f"{name}_predopt", predopt=True)
+
 
 torch.cuda.empty_cache()
 
-import os
 
 from tqdm.notebook import tqdm
 
 mae = nn.L1Loss()
 
 
-def exc_loss(reaction, pred_constants, dft="PBE", true_constants=true_constants_PBE):
+def exc_loss(
+    reaction, pred_constants, predicted_local_energies,  dft="PBE", true_constants=true_constants_PBE,
+    val=False
+):
+
     HARTREE2KCAL = 627.5095
 
     # Turn backsplit indices into slices.
@@ -191,13 +305,7 @@ def exc_loss(reaction, pred_constants, dft="PBE", true_constants=true_constants_
     )
     n_molecules = len(indices)
 
-    # Initialize loss.
-    loss = torch.tensor(0.0, requires_grad=True).to(device)
-
-    # Calculate predicted local energies.
-    predicted_local_energies = get_local_energies(
-        reaction, pred_constants, device, rung=rung, dft=dft
-    )["Local_energies"]
+    loss = torch.zeros(1, requires_grad=True).to(device)
 
     # Split them into systems.
     predicted_local_energies = [
@@ -210,19 +318,21 @@ def exc_loss(reaction, pred_constants, dft="PBE", true_constants=true_constants_
     )["Local_energies"]
 
     # Split them into systems.
-    true_local_energies = [true_local_energies[start:stop] for start, stop in indices]
+    true_local_energies = [
+        true_local_energies[start:stop] for start, stop in indices
+    ]
 
-    # Calculate local energy loss.
     for i in range(n_molecules):
         loss += (
-            1
-            / len(predicted_local_energies[i])
-            * torch.sqrt(
-                torch.sum((predicted_local_energies[i] - true_local_energies[i]) ** 2)
-            )
+            torch.sqrt(criterion(predicted_local_energies[i], true_local_energies[i]) + 1e-20)/np.sqrt(len(predicted_local_energies[i]))
         )
 
+    del (
+        true_local_energies,
+        predicted_local_energies
+    )
     return loss * HARTREE2KCAL / n_molecules
+
 
 true_constants_PBE = true_constants_PBE.to(device)
 
@@ -244,9 +354,15 @@ def train(
     test_loss_mae = []
     test_loss_mse = []
     test_loss_exc = []
+    train_full_loss = []
+    val_full_loss = []
+    test_fchem = []
+    train_fchem_loss = []
+
+    prev = None
 
     for epoch in range(n_epochs):
-        torch.autograd.set_detect_anomaly(True)
+        torch.autograd.set_detect_anomaly(False)
         print(f"Epoch {epoch + 1}")
         # train
         model.train()
@@ -254,63 +370,79 @@ def train(
         train_mae_losses_per_epoch = []
         train_mse_losses_per_epoch = []
         train_exc_losses_per_epoch = []
+        train_full_loss_per_epoch = []
         optimizer.zero_grad()
+
+        pred_energies = []
+        ref_energies = []
+        bases = []
+        errors = []
+        total_database_errors = {}
+
         for batch_idx, (X_batch, y_batch) in enumerate(progress_bar_train):
             X_batch_grid, y_batch = X_batch["Grid"].to(device), y_batch.to(device)
 
+            current_bases, bases = extend_bases(X_batch=X_batch, bases=bases)
+
             predictions = model(X_batch_grid)
-            reaction_energy = calculate_reaction_energy(
+            reaction_energy, local_energies = calculate_reaction_energy(
                 X_batch,
                 predictions,
                 device,
                 rung=rung,
                 dft=dft,
                 dispersions=dispersions,
-            ).to(device)
+            )
+            reaction_energy.to(device)
+            local_energies.to(device)
 
-            local_loss = exc_loss(X_batch, predictions, dft=dft)
-            reaction_mse_loss = criterion(reaction_energy, y_batch)
+            pred_energies, ref_energies, errors, total_database_errors = make_total_db_errors(
+                pred_energies, reaction_energy, errors, ref_energies, y_batch, total_database_errors, current_bases
+            )
 
-            # Calculate total loss function
-            loss = (1 - omega) / 5 * torch.sqrt(
-                reaction_mse_loss
-            ) + omega * local_loss * 100
-            MSE = reaction_mse_loss.item()
+            local_loss = exc_loss(X_batch, predictions, local_energies, dft=dft)
+
             MAE = mae(reaction_energy, y_batch).item()
 
+            batch_fchem_loss = batch_fchem(current_bases, reaction_energy, y_batch)
+
+            # Calculate total loss function
+            loss = (1 - omega) / 2  * batch_fchem_loss + omega * local_loss * 50
+            loss.backward()
+
+            MSE = criterion(reaction_energy, y_batch).item()
             train_mse_losses_per_epoch.append(MSE)
             train_mae_losses_per_epoch.append(MAE)
             train_exc_losses_per_epoch.append(local_loss.item())
+            train_full_loss_per_epoch.append(loss.item())
+            
             progress_bar_train.set_postfix(MSE=MSE, MAE=MAE)
 
-#            if verbose and epoch>1:
-#                if MAE/train_loss_mae[-1]>0.2:
-#                    print('TRAIN error significantly increased:\n')
-#                    print(
-#                        X_batch['Components']
-#                    )
-#                    print('Pred    True    Error')
-#                    for pred, true in zip(reaction_energy, y_batch):
-#                        print(f'{pred:7.2f} {true:7.2f} {abs(pred-true):7.2f}')
-
-            loss.backward()
             if ((batch_idx + 1) % accum_iter == 0) or (
                 batch_idx + 1 == len(train_loader)
             ):
                 optimizer.step()
                 optimizer.zero_grad()
 
-            del X_batch, X_batch_grid, y_batch, predictions, reaction_energy
+            del X_batch, X_batch_grid, y_batch, reaction_energy, local_energies, local_loss, MSE, MAE, loss, predictions
             gc.collect()
             torch.cuda.empty_cache()
 
         train_loss_mse.append(np.mean(train_mse_losses_per_epoch))
         train_loss_mae.append(np.mean(train_mae_losses_per_epoch))
         train_loss_exc.append(np.mean(train_exc_losses_per_epoch))
+        train_full_loss.append(np.mean(train_full_loss_per_epoch))
         print(
             f"train MSE Loss = {train_loss_mse[epoch]:.8f} MAE Loss = {train_loss_mae[epoch]:.8f}"
         )
-        print(f"train Local Energy Loss = {train_loss_exc[epoch]:.8f}")
+        print(
+            f"train Local Energy Loss = {train_loss_exc[epoch]:.8f}"
+        )
+
+        train_fchem = loss_function(factor_dictionary=FCHEM_VALIDATION, total_database_errors=total_database_errors)
+        print("\nTrain Fchem", train_fchem, '\n')
+        train_fchem_loss.append(train_fchem)
+
 
         # test
         model.eval()
@@ -318,49 +450,66 @@ def train(
         test_mae_losses_per_epoch = []
         test_mse_losses_per_epoch = []
         test_exc_losses_per_epoch = []
+        val_full_loss_per_epoch = []
+
+        pred_energies = []
+        ref_energies = []
+        bases = []
+        errors = []
+        total_database_errors = {}
+
+
         with torch.no_grad():
             for X_batch, y_batch in progress_bar_test:
                 X_batch_grid, y_batch = X_batch["Grid"].to(device), y_batch.to(device)
 
-                X_batch_grid = torch.tanh(torch.exp(X_batch_grid*STD_TRAIN_FEATURES+MEAN_TRAIN_FEATURES)-1e-8)
+                current_bases, bases = extend_bases(X_batch=X_batch, bases=bases)
 
                 predictions = model(X_batch_grid)
-                local_loss = exc_loss(X_batch, predictions, dft=dft)
-                reaction_energy = calculate_reaction_energy(
+                reaction_energy, local_energies = calculate_reaction_energy(
                     X_batch,
                     predictions,
                     device,
                     rung=rung,
                     dft=dft,
                     dispersions=dispersions,
-                ).to(device)
+                )
+                reaction_energy.to(device)
+                local_energies.to(device)
 
-                loss = criterion(reaction_energy, y_batch)
-                MSE = loss.item()
+
+                pred_energies, ref_energies, errors, total_database_errors = make_total_db_errors(
+                    pred_energies, reaction_energy, errors, ref_energies, y_batch, total_database_errors, current_bases
+                )
+
+                local_loss = exc_loss(X_batch, predictions, local_energies, dft=dft)
+
                 MAE = mae(reaction_energy, y_batch).item()
+
+                reaction_mse_loss = criterion(reaction_energy, y_batch)
+                batch_fchem_loss = batch_fchem(current_bases, reaction_energy, y_batch)
+
+                MSE = reaction_mse_loss.item()
+
+                loss = (1 - omega) / 2 * batch_fchem_loss + omega * local_loss * 50
+
                 test_mse_losses_per_epoch.append(MSE)
                 test_mae_losses_per_epoch.append(MAE)
                 test_exc_losses_per_epoch.append(local_loss.item())
+                val_full_loss_per_epoch.append(loss.item())
 
-                if verbose and epoch>1:
-                    if MAE/test_loss_mae[-1]>1.2:
-                        print('VALIDATION error significantly increased:\n')
-                        print(
-                            X_batch['Components']
-                        )
-                        print('Pred    True    Error')
-                        for pred, true in zip(reaction_energy, y_batch):
-                            print(f'{pred:7.2f} {true:7.2f} {abs(pred-true):7.2f}')
                 progress_bar_test.set_postfix(MSE=MSE, MAE=MAE)
                 del (
                     X_batch,
                     X_batch_grid,
                     y_batch,
-                    predictions,
                     reaction_energy,
                     loss,
                     MAE,
                     MSE,
+                    local_energies,
+                    local_loss,
+                    predictions
                 )
                 gc.collect()
                 torch.cuda.empty_cache()
@@ -368,20 +517,45 @@ def train(
         test_loss_mse.append(np.mean(test_mse_losses_per_epoch))
         test_loss_mae.append(np.mean(test_mae_losses_per_epoch))
         test_loss_exc.append(np.mean(test_exc_losses_per_epoch))
-
+        val_full_loss.append(np.mean(val_full_loss_per_epoch))
+        if scheduler:
+            scheduler.step()
         print(
             f"test MSE Loss = {test_loss_mse[epoch]:.8f} MAE Loss = {test_loss_mae[epoch]:.8f}"
         )
+        fig, ax = plt.subplots(nrows=2, ncols=1, figsize=[3, 6], sharex=True)
+        ax[0].plot(range(1, len(train_full_loss)+1), train_full_loss, label='Train Loss')
+        ax[0].plot(range(1, len(val_full_loss)+1), val_full_loss, label='Validation Loss')
         print(f"test Local Energy Loss = {test_loss_exc[epoch]:.8f}")
 
-        # Save the model every 10 epochs.
-        if (epoch + 1) % 10 == 0:
-            log_params(
-                model,
-                train_loss_mae,
-                test_loss_mae,
-                name=f"{name}_train_epoch_{epoch+1}",
+        val_fchem = loss_function(factor_dictionary=FCHEM_VALIDATION, total_database_errors=total_database_errors, val=True)
+        test_fchem.append(val_fchem)
+
+        if val_fchem <= min(test_fchem):
+            if prev:
+                os.remove(prev)
+            prev = f'best_models/RELATIVE_AE17/lr_{lr_train}_{name}_{omega}_epoch_{epoch+1}_loss_{val_fchem:.2f}_{test_loss_exc[-1]:.4f}.pth'
+            torch.save(
+                model.state_dict(), prev
             )
+        print("\nValidation Fchem", val_fchem, '\n')
+        ax[1].plot(range(1, len(val_full_loss)+1), train_fchem_loss, label='Train Fchem')
+        ax[1].plot(range(1, len(val_full_loss)+1), test_fchem, label='Validation Fchem')
+        ax[0].legend()
+        ax[1].legend()
+        plt.savefig(f"./batch_fchem/RELATIVE_AE17/lr_{lr_train}_{name}_{omega}.png")
+        plt.clf()
+        plt.close()
+
+        del (
+            val_fchem,
+            train_fchem,
+            test_mse_losses_per_epoch,
+            test_mae_losses_per_epoch,
+            val_full_loss_per_epoch
+        )
+        gc.collect()
+        torch.cuda.empty_cache()
 
     return train_loss_mae, test_loss_mae
 
@@ -398,11 +572,17 @@ reload(utils)
 reload(reaction_energy_calculation)
 reload(PBE)
 
-optimizer = torch.optim.RMSprop(model.parameters(), lr=lr_train)
+optimizer = torch.optim.Adam(
+    model.parameters(), lr=lr_train, betas=(0.9, 0.999), weight_decay=0.01
+)
+if dft=='XALPHA':
+    scheduler = StepLR(optimizer, step_size=100, gamma=0.1)
+else:
+    scheduler = None
 
 N_EPOCHS = n_train
 ACCUM_ITER = 1
-VERBOSE = True
+VERBOSE = False
 train_loss_mae, test_loss_mae = train(
     model,
     criterion,
@@ -414,6 +594,3 @@ train_loss_mae, test_loss_mae = train(
     omega=omega,
     verbose=VERBOSE
 )
-
-# Log the final model.
-log_params(model, train_loss_mae, test_loss_mae, name=f"{name}_final")
